fix: use decimal instead of double and use numeric instead of real

+++ src/backend/Application/Features/WeightEntryFeatures/CreateWeightEntry/CreateWeightEntryRequest.cs
@@ -2,7 +2,7 @@ namespace Application.Features.WeightEntryFeatures.CreateWeightEntry;
 
 public class CreateWeightEntryRequest
 {
-    public required double Value { get; set; }
+    public required decimal Value { get; set; }
     public string? Comment { get; set; }
     public required DateOnly EntryDate { get; set; }
 }
